python/requirements.txt
requests
apscheduler
mysql-connector-python